splttr-api/requirements.txt
Django==1.9.7
django-rest-swagger==0.3.8
djangorestframework==3.3.3
Markdown==2.6.6
psycopg2==2.6.2
PyYAML==3.11
